Fix routings
