+++ frontend/src/pages/Home/Login.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-const Home = () => {
+const Login = () => {
   return (
     <>
       <div>
@@ -10,4 +10,4 @@ const Home = () => {
   );
 };
 
-export default Home;
+export default Login;
